+++ modules/db_mysql_aurora/iam.tf
@@ -1,0 +1,31 @@
+
+# IAM Role for Enhanced Monitoring (only when needed)
+# https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/iam_role
+resource "aws_iam_role" "enhanced_monitoring" {
+  count = local.enhanced_monitoring_interval > 0 ? 1 : 0
+
+  name_prefix = "aurora-mysql-${var.name}-monitoring-"
+
+  assume_role_policy = jsonencode({
+    Version = "2012-10-17"
+    Statement = [
+      {
+        Action = "sts:AssumeRole"
+        Effect = "Allow"
+        Principal = {
+          Service = "monitoring.rds.amazonaws.com"
+        }
+      }
+    ]
+  })
+
+  tags = var.tags
+}
+
+# https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/iam_role_policy_attachment
+resource "aws_iam_role_policy_attachment" "enhanced_monitoring" {
+  count = local.enhanced_monitoring_interval > 0 ? 1 : 0
+
+  role       = aws_iam_role.enhanced_monitoring[0].name
+  policy_arn = "arn:aws:iam::aws:policy/service-role/AmazonRDSEnhancedMonitoringRole"
+}