+++ modules/db_mysql_aurora/providers.tf
@@ -1,0 +1,11 @@
+
+terraform {
+  required_version = ">= 1.10"
+  required_providers {
+    # https://registry.terraform.io/providers/hashicorp/aws/latest
+    aws = {
+      source  = "hashicorp/aws"
+      version = "6.10.0"
+    }
+  }
+}